rename package and files: characterList -> charactersList
