+++ ui-mobile/src/main/java/com/choala/rickandmortyallinfo/charactersList/CharactersListAdapter.kt
@@ -1,4 +1,4 @@
-package com.choala.rickandmortyallinfo.characterList
+package com.choala.rickandmortyallinfo.charactersList
 
 import android.view.ViewGroup
 import androidx.paging.PagingDataAdapter
